test: add list areas and list spaces tests
