+++ src/test/java/br/edu/ifsp/spo/eventos/eventplatformbackend/e2e/location/DeleteLocationApiTest.java
@@ -1,5 +1,4 @@
-package br.edu.ifsp.spo.eventos.eventplatformbackend.e2e.space;
-
+package br.edu.ifsp.spo.eventos.eventplatformbackend.e2e.location;
 import io.restassured.http.ContentType;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.DisplayName;
@@ -10,36 +9,30 @@ import org.springframework.test.context.jdbc.Sql;
 
 import static io.restassured.RestAssured.*;
 import static org.hamcrest.Matchers.*;
-import static org.hamcrest.Matchers.equalTo;
 
 @SpringBootTest(webEnvironment = SpringBootTest.WebEnvironment.RANDOM_PORT)
-public class DeleteSpaceApiTest {
+public class DeleteLocationApiTest {
     @LocalServerPort
     private int localPort;
-    private String spaceURI;
+    private String locationURI;
 
     @BeforeEach
     public void beforeEach() {
         baseURI = "http://localhost";
         port = localPort;
-        spaceURI = "/api/v1/locations/{locationId}/areas/{areaId}/spaces/{spaceId}";
+        locationURI = "/api/v1/locations/{locationId}";
     }
-
     @Test
-    @DisplayName("DELETE /locations/{locationId}/areas/{areaId}/spaces/{spaceId}")
+    @DisplayName("DELETE /locations/{locationId}")
     @Sql("/sql/delete_all_tables.sql")
-    @Sql("/sql/locations/insert_one.sql")
-    @Sql("/sql/areas/insert_one.sql")
-    @Sql("/sql/spaces/insert_one.sql")
-    public void deleteSpace() {
+    @Sql("/sql/locations/insert_many.sql")
+    public void deleteLocation() {
         given()
             .contentType(ContentType.JSON)
             .pathParam("locationId", "5607ddd3-31ed-4435-bd61-23133d2f3381")
-            .pathParam("areaId", "29eb3ccf-711d-40f2-954c-3f2616a6cf36")
-            .pathParam("spaceId", "a7118369-b18b-48e9-b172-ba23be91d9d5")
             .log().all()
         .when()
-            .delete(spaceURI)
+            .delete(locationURI)
         .then()
             .log().all()
             .assertThat()
@@ -47,25 +40,41 @@ public class DeleteSpaceApiTest {
     }
 
     @Test
-    @DisplayName("DELETE /locations/{locationId}/areas/{areaId}/spaces/{spaceId} - not found")
+    @DisplayName("DELETE /locations/{locationId} - not found")
     @Sql("/sql/delete_all_tables.sql")
-    @Sql("/sql/locations/insert_one.sql")
-    @Sql("/sql/areas/insert_one.sql")
-    @Sql("/sql/spaces/insert_one.sql")
-    public void deleteSpaceNotFound() {
+    @Sql("/sql/locations/insert_many.sql")
+    public void deleteLocationNotFound() {
         given()
             .contentType(ContentType.JSON)
-            .pathParam("locationId", "5607ddd3-31ed-4435-bd61-23133d2f3381")
-            .pathParam("areaId", "29eb3ccf-711d-40f2-954c-3f2616a6cf36")
-            .pathParam("spaceId", "a6118369-b18b-48e9-b172-ba23be91d9d5")
+            .pathParam("locationId", "a7118369-b18b-48e9-b172-ba23be91d9d5")
             .log().all()
         .when()
-            .delete(spaceURI)
+            .delete(locationURI)
         .then()
             .log().all()
             .assertThat()
                 .statusCode(404)
                 .body("title", equalTo("Resource not found exception"))
                 .body("violations", hasSize(1));
+    }
+
+    @Test
+    @DisplayName("DELETE /locations/{locationId} - with areas associated")
+    @Sql("/sql/delete_all_tables.sql")
+    @Sql("/sql/locations/insert_one.sql")
+    @Sql("/sql/areas/insert_one.sql")
+    public void deleteLocationWithAreasAssociated() {
+        given()
+            .contentType(ContentType.JSON)
+            .pathParam("locationId", "5607ddd3-31ed-4435-bd61-23133d2f3381")
+            .log().all()
+        .when()
+            .delete(locationURI)
+        .then()
+            .log().all()
+            .assertThat()
+                .statusCode(409)
+                .body("title", equalTo("Resource referential integrity exception"))
+                .body("violations", hasSize(2));
     }
 }
