--- src/test/java/br/edu/ifsp/spo/eventos/eventplatformbackend/e2e/location/LocationApiTest.java
@@ -1,334 +1,0 @@
-package br.edu.ifsp.spo.eventos.eventplatformbackend.e2e.location;
-
-import io.restassured.http.ContentType;
-import org.junit.jupiter.api.AfterEach;
-import org.junit.jupiter.api.BeforeEach;
-import org.junit.jupiter.api.DisplayName;
-import org.junit.jupiter.api.Test;
-import org.springframework.boot.test.context.SpringBootTest;
-import org.springframework.boot.test.web.server.LocalServerPort;
-import org.springframework.test.context.jdbc.Sql;
-
-import static io.restassured.RestAssured.*;
-import static org.hamcrest.Matchers.*;
-
-
-// https://www.javadoc.io/doc/io.rest-assured/rest-assured/4.5.1/index.html
-// https://github.com/rest-assured/rest-assured/wiki/Usage
-@SpringBootTest(webEnvironment = SpringBootTest.WebEnvironment.RANDOM_PORT)
-public class LocationApiTest {
-    @LocalServerPort
-    private int localPort;
-    private String locationsURI;
-    private String locationURI;
-
-    @BeforeEach
-    public void beforeEach() {
-        baseURI = "http://localhost";
-        port = localPort;
-        locationsURI = "/api/v1/locations";
-        locationURI = "/api/v1/locations/{locationId}";
-    }
-
-    @Test
-    @DisplayName("GET /locations - empty list")
-    @Sql("/sql/delete_all_tables.sql")
-    public void getAllEmptyList() {
-        given()
-            .contentType(ContentType.JSON)
-            .log().all()
-        .when()
-            .get(locationsURI)
-        .then()
-            .log().all()
-            .assertThat()
-                .statusCode(200)
-                .body("$", hasSize(0));
-    }
-
-    @Test
-    @DisplayName("GET /locations - one location inside list")
-    @Sql("/sql/delete_all_tables.sql")
-    @Sql("/sql/locations/insert_one.sql")
-    public void getListOneLocations() {
-        given()
-            .contentType(ContentType.JSON)
-            .log().all()
-        .when()
-            .get(locationsURI)
-        .then()
-            .log().all()
-            .assertThat()
-                .statusCode(200)
-                .body("$", hasSize(1));
-    }
-
-    @Test
-    @DisplayName("GET /locations - multiples location inside list")
-    @Sql("/sql/delete_all_tables.sql")
-    @Sql("/sql/locations/insert_many.sql")
-    public void getListMultiplesLocations() {
-        given()
-            .contentType(ContentType.JSON)
-            .log().all()
-        .when()
-            .get(locationsURI)
-        .then()
-            .log().all()
-            .assertThat()
-                .statusCode(200)
-                .body("$", hasSize(3));
-    }
-
-    @Test
-    @DisplayName("GET /locations/{id}")
-    @Sql("/sql/delete_all_tables.sql")
-    @Sql("/sql/locations/insert_many.sql")
-    public void getLocation() {
-        given()
-            .contentType(ContentType.JSON)
-            .pathParam("locationId", "5607ddd3-31ed-4435-bd61-23133d2f3381")
-            .log().all()
-        .when()
-            .get(locationURI)
-        .then()
-            .log().all()
-            .assertThat()
-                .statusCode(200)
-                .body("id", notNullValue())
-                .body("name", equalTo("IFSP Campus São Paulo"))
-                .body("address", equalTo("Rua Pedro Vicente, 625 - Canindé, São Paulo - SP, 01109-010"));
-    }
-
-    @Test
-    @DisplayName("GET /locations/{locationId} - not found")
-    @Sql("/sql/delete_all_tables.sql")
-    @Sql("/sql/locations/insert_many.sql")
-    public void getLocationNotFound() {
-        given()
-            .contentType(ContentType.JSON)
-            .pathParam("locationId", "a7118369-b18b-48e9-b172-ba23be91d9d5")
-            .log().all()
-        .when()
-            .get(locationURI)
-        .then()
-            .log().all()
-            .assertThat()
-                .statusCode(404)
-                .body("title", equalTo("Resource not found exception"))
-                .body("violations", hasSize(1));
-    }
-
-    @Test
-    @DisplayName("POST /locations")
-    @Sql("/sql/delete_all_tables.sql")
-    public void postLocations() {
-        given()
-            .contentType(ContentType.JSON)
-            .body("""
-                {
-                   "name": "IFSP Campus São Paulo",
-                   "address": "Rua Pedro Vicente, 625 - Canindé, São Paulo - SP, 01109-010"
-                }
-            """)
-            .log().all()
-        .when()
-            .post(locationsURI)
-        .then()
-            .log().all()
-            .assertThat()
-                .statusCode(201)
-                .body("id", notNullValue())
-                .body("name", equalTo("IFSP Campus São Paulo"))
-                .body("address", equalTo("Rua Pedro Vicente, 625 - Canindé, São Paulo - SP, 01109-010"));
-
-    }
-
-    @Test
-    @DisplayName("POST /locations - already exists")
-    @Sql("/sql/delete_all_tables.sql")
-    @Sql("/sql/locations/insert_many.sql")
-    public void postLocationsAlreadyExists() {
-        given()
-            .contentType(ContentType.JSON)
-            .body("""
-                {
-                   "name": "IFSP Campus São Paulo",
-                   "address": "Rua Pedro Vicente, 625 - Canindé, São Paulo - SP, 01109-010"
-                }
-            """)
-            .log().all()
-        .when()
-            .post(locationsURI)
-        .then()
-            .log().all()
-            .assertThat()
-                .statusCode(409)
-                .body("title", equalTo("Resource already exists exception"))
-                .body("violations", hasSize(1));
-
-    }
-
-    @Test
-    @DisplayName("PUT /locations/{locationId}")
-    @Sql("/sql/delete_all_tables.sql")
-    @Sql("/sql/locations/insert_many.sql")
-    public void putLocations() {
-        given()
-            .contentType(ContentType.JSON)
-            .pathParam("locationId", "5607ddd3-31ed-4435-bd61-23133d2f3381")
-            .body("""
-                {
-                   "name": "IFSP Campus SP",
-                   "address": "Rua Pedro Vicente, 100 - Canindé, São Paulo - SP, 01109-010"
-                }
-            """)
-            .log().all()
-        .when()
-            .put(locationURI)
-        .then()
-            .log().all()
-            .assertThat()
-                .statusCode(200)
-                .body("id", notNullValue())
-                .body("name", equalTo("IFSP Campus SP"))
-                .body("address", equalTo("Rua Pedro Vicente, 100 - Canindé, São Paulo - SP, 01109-010"));
-
-    }
-
-    @Test
-    @DisplayName("PUT /locations/{locationId}")
-    @Sql("/sql/delete_all_tables.sql")
-    @Sql("/sql/locations/insert_many.sql")
-    public void putLocationsChangeAddressOnly() {
-        given()
-            .contentType(ContentType.JSON)
-            .pathParam("locationId", "5607ddd3-31ed-4435-bd61-23133d2f3381")
-            .body("""
-                {
-                   "name": "IFSP Campus São Paulo",
-                   "address": "Rua Pedro Vicente, 100 - Canindé, São Paulo - SP, 01109-010"
-                }
-            """)
-            .log().all()
-        .when()
-            .put(locationURI)
-        .then()
-            .log().all()
-            .assertThat()
-                .statusCode(200)
-                .body("id", notNullValue())
-                .body("name", equalTo("IFSP Campus São Paulo"))
-                .body("address", equalTo("Rua Pedro Vicente, 100 - Canindé, São Paulo - SP, 01109-010"));
-
-    }
-
-
-    @Test
-    @DisplayName("PUT /locations/{locationId} - already exists")
-    @Sql("/sql/delete_all_tables.sql")
-    @Sql("/sql/locations/insert_many.sql")
-    public void putLocationsAlreadyExists() {
-        given()
-            .contentType(ContentType.JSON)
-            .pathParam("locationId", "5607ddd3-31ed-4435-bd61-23133d2f3381")
-            .body("""
-                {
-                   "name": "Shopping D",
-                   "address": "Rua Pedro Vicente, 100 - Canindé, São Paulo - SP, 01109-010"
-                }
-            """)
-            .log().all()
-        .when()
-            .put(locationURI)
-        .then()
-            .log().all()
-            .assertThat()
-                .statusCode(409)
-                .body("title", equalTo("Resource already exists exception"))
-                .body("violations", hasSize(1));
-
-    }
-
-    @Test
-    @DisplayName("PUT /locations/{locationId} - not found")
-    @Sql("/sql/delete_all_tables.sql")
-    @Sql("/sql/locations/insert_one.sql")
-    public void putLocationsNotFound() {
-        given()
-            .contentType(ContentType.JSON)
-            .pathParam("locationId", "2e07b7bd-b4ab-46fc-b19c-b3425b668b3b")
-            .body("""
-                {
-                   "name": "Shopping D",
-                   "address": "Rua Pedro Vicente, 100 - Canindé, São Paulo - SP, 01109-010"
-                }
-            """)
-            .log().all()
-        .when()
-            .put(locationURI)
-        .then()
-            .log().all()
-            .assertThat()
-                .statusCode(404)
-                .body("title", equalTo("Resource not found exception"))
-                .body("violations", hasSize(1));
-
-    }
-
-    @Test
-    @DisplayName("DELETE /locations/{id}")
-    @Sql("/sql/delete_all_tables.sql")
-    @Sql("/sql/locations/insert_many.sql")
-    public void deleteLocation() {
-        given()
-            .contentType(ContentType.JSON)
-            .pathParam("locationId", "5607ddd3-31ed-4435-bd61-23133d2f3381")
-            .log().all()
-        .when()
-            .delete(locationURI)
-        .then()
-            .log().all()
-            .assertThat()
-                .statusCode(204);
-    }
-
-    @Test
-    @DisplayName("DELETE /locations/{locationId} - not found")
-    @Sql("/sql/delete_all_tables.sql")
-    @Sql("/sql/locations/insert_many.sql")
-    public void deleteLocationNotFound() {
-        given()
-            .contentType(ContentType.JSON)
-            .pathParam("locationId", "a7118369-b18b-48e9-b172-ba23be91d9d5")
-            .log().all()
-        .when()
-            .delete(locationURI)
-        .then()
-            .log().all()
-            .assertThat()
-            .statusCode(404)
-                .body("title", equalTo("Resource not found exception"))
-                .body("violations", hasSize(1));
-    }
-
-    @Test
-    @DisplayName("DELETE /locations/{locationId} - with areas associated")
-    @Sql("/sql/delete_all_tables.sql")
-    @Sql("/sql/locations/insert_one.sql")
-    @Sql("/sql/areas/insert_one.sql")
-    public void deleteLocationWithAreasAssociated() {
-        given()
-            .contentType(ContentType.JSON)
-            .pathParam("locationId", "5607ddd3-31ed-4435-bd61-23133d2f3381")
-            .log().all()
-        .when()
-            .delete(locationURI)
-            .then()
-        .log().all()
-            .assertThat()
-            .statusCode(409)
-                .body("title", equalTo("Resource referential integrity exception"))
-                .body("violations", hasSize(2));
-    }
-}
