+++ src/test/java/br/edu/ifsp/spo/eventos/eventplatformbackend/e2e/space/SpaceDetailsApiTest.java
@@ -64,16 +64,16 @@ public class SpaceDetailsApiTest {
     @Sql("/sql/spaces/insert_one.sql")
     public void getSpaceNotFound() {
         given()
-                .contentType(ContentType.JSON)
-                .pathParam("locationId", "5607ddd3-31ed-4435-bd61-23133d2f3381")
-                .pathParam("areaId", "29eb3ccf-711d-40f2-954c-3f2616a6cf36")
-                .pathParam("spaceId", "a6118369-b18b-48e9-b172-ba23be91d9d5")
-                .log().all()
-                .when()
-                .get(spaceURI)
-                .then()
-                .log().all()
-                .assertThat()
+            .contentType(ContentType.JSON)
+            .pathParam("locationId", "5607ddd3-31ed-4435-bd61-23133d2f3381")
+            .pathParam("areaId", "29eb3ccf-711d-40f2-954c-3f2616a6cf36")
+            .pathParam("spaceId", "a6118369-b18b-48e9-b172-ba23be91d9d5")
+            .log().all()
+        .when()
+            .get(spaceURI)
+        .then()
+            .log().all()
+            .assertThat()
                 .statusCode(404)
                 .body("title", equalTo("Resource not found exception"))
                 .body("violations", hasSize(1));
